+++ src/Writing/HtmlWriter.php
@@ -187,7 +187,7 @@ class HtmlWriter
             }
         }
 
-        $headings = array_merge($headings, array_map(function ($group) {
+        $headings = array_merge($headings, array_values(array_map(function ($group) {
             $groupSlug = Str::slug($group['name']);
 
             return [
@@ -199,7 +199,7 @@ class HtmlWriter
                             'slug' => $endpoint->fullSlug(),
                             'name' => $endpoint->name(),
                             'subheadings' => []
-                        ])->all();
+                        ])->values();
                     }
 
                     return [
@@ -210,12 +210,12 @@ class HtmlWriter
                                 'slug' => $endpoint->fullSlug(),
                                 'name' => $endpoint->name(),
                                 'subheadings' => []
-                            ])->all(),
+                            ])->values(),
                         ],
                     ];
-                })->all(),
+                })->values(),
             ];
-        }, $endpointsByGroupAndSubgroup));
+        }, $endpointsByGroupAndSubgroup)));
 
         $lastL1ElementIndex = null;
         foreach ($headingsAfterEndpoints as $heading) {