+++ src/Writing/OpenAPISpecWriter.php
@@ -19,7 +19,7 @@ class OpenAPISpecWriter
 {
     use ParamHelpers;
 
-    const VERSION = '3.0.3';
+    const SPEC_VERSION = '3.0.3';
 
     private DocumentationConfig $config;
 
@@ -46,7 +46,7 @@ class OpenAPISpecWriter
     public function generateSpecContent(array $groupedEndpoints): array
     {
         return array_merge([
-            'openapi' => self::VERSION,
+            'openapi' => self::SPEC_VERSION,
             'info' => [
                 'title' => $this->config->get('title') ?: config('app.name', ''),
                 'description' => $this->config->get('description', ''),
@@ -84,6 +84,7 @@ class OpenAPISpecWriter
             $operations = $endpoints->mapWithKeys(function (OutputEndpointData $endpoint) use ($groupedEndpoints) {
                 $spec = [
                     'summary' => $endpoint->metadata->title,
+                    'operationId' => $this->operationId($endpoint),
                     'description' => $endpoint->metadata->description,
                     'parameters' => $this->generateEndpointParametersSpec($endpoint),
                     'responses' => $this->generateEndpointResponsesSpec($endpoint),
@@ -534,5 +535,13 @@ class OpenAPISpecWriter
                 'example' => $field->example,
             ];
         }
+    }
+
+    function operationId(OutputEndpointData $endpoint): string
+    {
+        if ($endpoint->metadata->title) return preg_replace('/[^\w+]/', '', Str::camel($endpoint->metadata->title));
+
+        $parts = preg_split('/[^\w+]/', $endpoint->uri, -1, PREG_SPLIT_NO_EMPTY);
+        return Str::lower($endpoint->httpMethods[0]) . join('', array_map(fn ($part) => ucfirst($part), $parts));
     }
 }