+++ tests/Unit/ExtractedEndpointDataTest.php
@@ -6,6 +6,7 @@ use Illuminate\Routing\Route as LaravelRoute;
 use Illuminate\Support\Facades\Route;
 use Knuckles\Camel\Extraction\ExtractedEndpointData;
 use Knuckles\Scribe\Matching\RouteMatcher;
+use Knuckles\Scribe\Scribe;
 use Knuckles\Scribe\Tests\BaseLaravelTest;
 use Knuckles\Scribe\Tests\Fixtures\TestController;
 
@@ -29,6 +30,28 @@ class ExtractedEndpointDataTest extends BaseLaravelTest
     }
 
     /** @test */
+    public function allows_user_specified_normalization()
+    {
+        Scribe::normalizeEndpointUrlUsing(function (string $url, LaravelRoute $route, \ReflectionFunctionAbstract $method, ?\ReflectionClass $controller) {
+            if ($url == 'things/{thing}') return 'things/{the_id_of_the_thing}';
+
+            if ($route->named('things.otherthings.destroy')) return 'things/{thing-id}/otherthings/{other_thing-id}';
+        });
+
+        Route::apiResource('things', TestController::class)->only('show');
+        $route = $this->getRoute(['prefixes' => '*']);
+        $this->assertEquals('things/{thing}', $this->originalUri($route));
+        $this->assertEquals('things/{the_id_of_the_thing}', $this->expectedUri($route));
+
+        Route::apiResource('things.otherthings', TestController::class)->only('destroy');
+        $route = $this->getRoute(['prefixes' => '*/otherthings/*']);
+        $this->assertEquals('things/{thing}/otherthings/{otherthing}', $this->originalUri($route));
+        $this->assertEquals('things/{thing-id}/otherthings/{other_thing-id}', $this->expectedUri($route));
+
+        Scribe::normalizeEndpointUrlUsing(null);
+    }
+
+    /** @test */
     public function normalizes_resource_url_params_from_underscores_to_hyphens()
     {
         Route::apiResource('audio-things', TestController::class)->only('show');
@@ -36,6 +59,12 @@ class ExtractedEndpointDataTest extends BaseLaravelTest
 
         $this->assertEquals('audio-things/{audio_thing}', $this->originalUri($route));
         $this->assertEquals('audio-things/{id}', $this->expectedUri($route));
+
+        Route::apiResource('big-users.audio-things.things', TestController::class)->only('store');
+        $route = $this->getRoute(['prefixes' => '*big-users*']);
+
+        $this->assertEquals('big-users/{big_user}/audio-things/{audio_thing}/things', $this->originalUri($route));
+        $this->assertEquals('big-users/{big_user_id}/audio-things/{audio_thing_id}/things', $this->expectedUri($route));
     }
 
     /** @test */